+++ CHANGES.md
@@ -1,3 +1,10 @@
+### 2.4.0
+
+* Experimental support for [vert.x 2.x](http://vertx.io). Should now run in vert.x >= 1.1.0.
+* New `when.isPromiseLike` as the more accurately-named synonym for `when.isPromise`.
+* **DEPRECATED**: `when.isPromise`. It can only tell you that something is "promise-like" (aka "thenable") anyway. Use the new, more accurately-named `when.isPromiseLike` instead.
+* Fix for promise monitor reporting extra unhandled rejections for `when.all` and `when.map`.
+
 ### 2.3.0
 
 * New [`promise.tap`](docs/api.md#tap) for adding side effects to a promise chain.